SWITCHER-81 Se elimino la carpeta con los fetches pues no funcionaba, se termino utilizando el fetch en la misma funcion donde se llama
